fix rebase damage, move keystore + txm to pkg/chainlink
