+++ relayer/pkg/chainlink/txm/txm_test.go
@@ -8,10 +8,10 @@ import (
 	"github.com/dontpanicdao/caigo/types"
 	"github.com/pkg/errors"
 	"github.com/smartcontractkit/chainlink-relay/pkg/logger"
-	"github.com/smartcontractkit/chainlink-starknet/pkg/starknet"
-	"github.com/smartcontractkit/chainlink-starknet/pkg/starknet/db"
-	"github.com/smartcontractkit/chainlink-starknet/pkg/starknet/keys"
-	"github.com/smartcontractkit/chainlink-starknet/pkg/starknet/keys/mocks"
+	"github.com/smartcontractkit/chainlink-starknet/relayer/pkg/chainlink/keys"
+	"github.com/smartcontractkit/chainlink-starknet/relayer/pkg/chainlink/keys/mocks"
+	"github.com/smartcontractkit/chainlink-starknet/relayer/pkg/starknet"
+	"github.com/smartcontractkit/chainlink-starknet/relayer/pkg/starknet/db"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 )
